@@ -76,21 +68,33 @@
     "rimraf": "^2.6.2",
     "selenium-webdriver": "^4.0.0-alpha.1",
     "style-loader": "^0.21.0",
-    "webpack": "^4.12.0",
+    "webpack": "^4.29.6",
     "webpack-bundle-analyzer": "^3.0.3",
     "webpack-cli": "^3.0.7",
     "webpack-dev-server": "^3.1.4"
   },
   "dependencies": {
-    "leaflet.markercluster": "^1.4.1",
-    "react-leaflet-markercluster": "^2.0.0-rc3",
+    "object-path": "^0.11.4",
+    "@babel/plugin-syntax-dynamic-import": "^7.2.0",
+    "@babel/polyfill": "^7.4.3",
+    "@babel/register": "^7.4.0",
+    "@types/chart.js": "^2.7.50",
+    "@types/jest": "^24.0.11",
+    "@types/react": "^16.8.12",
+    "babel-eslint": "^11.0.0-beta.0",
+    "babel-loader": "^8.0.0-beta.6",
+    "babel-plugin-add-react-displayname": "^0.0.5",
+    "babel-plugin-module-resolver": "^3.2.0",
+    "babel-plugin-react-intl": "^3.0.1",
     "babel-polyfill": "^6.26.0",
-    "babel-watch": "^2.0.7",
-    "chart.js": "^2.7.2",
+    "babel-watch": "^7.0.0",
+    "chart.js": "^2.8.0",
     "concurrently": "^4.0.1",
+    "event-source-polyfill": "^1.0.5",
     "fullcalendar-reactwrapper": "^1.0.7",
     "jquery": "^3.3.1",
     "leaflet": "^1.4.0",
+    "leaflet.markercluster": "^1.4.1",
     "lodash": "^4.17.10",
     "mini-css-extract-plugin": "^0.4.4",
     "moment": "^2.22.2",
