@@ -1,5 +1,4 @@
-
 module.exports = {
   defineMessages: () => ({}),
-  FormattedMessage: () => 'Message'
+  FormattedMessage: () => "Message"
 }
