@@ -5,11 +5,18 @@ import Select2 from "aurigauikit/components/Select2"
 import { Grid, Div } from "aurigauikit/components/Grid"
 import messages from "./messages"
 
+const newId = { id: "dashboard-page.add-widget.new" }
+
 export default injectIntl(
   class extends React.Component {
     state = { selectedWidget: this.props.widgets[0] }
 
-    changeWidget = selectedWidget => this.setState({ selectedWidget })
+    changeWidget = selectedWidget => {
+      const { intl } = this.props
+      return selectedWidget.name === intl.formatMessage(newId)
+        ? this.setState({ selectedWidget: { name: intl.formatMessage(newId) } })
+        : this.setState({ selectedWidget })
+    }
 
     confirm = () => {
       const { onAdd } = this.props
@@ -17,21 +24,24 @@ export default injectIntl(
       onAdd(selectedWidget)
     }
 
+    getWidgets = () => {
+      const { widgets, onBuild, intl } = this.props
+      const widgetList = widgets.filter(({ name }) => name !== "active-assets")
+      return onBuild ? [...widgetList, { name: intl.formatMessage(newId) }] : widgetList
+    }
+
     render() {
-      const { onClose, widgets, messages: widgetsMessages, intl } = this.props
+      const { onClose, messages: widgetsMessages, intl } = this.props
       const { selectedWidget } = this.state
       return (
         <Modal show onClose={() => onClose()} style={{ width: "70%" }}>
-          <Modal.Header title={<Msg {...messages.modal.title} />}>
-            {/*<button className="btn btn-warning"> Edit </button>
-            <button className="btn btn-critical"> Remove </button>*/}
-          </Modal.Header>
+          <Modal.Header title={<Msg {...messages.modal.title} />} />
           <Modal.Content>
             <Grid padding={20}>
               <Div col="1-2">
                 <Select2
                   style={{ width: "100%" }}
-                  data={widgets}
+                  data={this.getWidgets()}
                   value={selectedWidget}
                   didSelect={this.changeWidget}
                   willDisplay={({ name }) =>
@@ -44,27 +54,37 @@ export default injectIntl(
                 />
               </Div>
             </Grid>
-            <div
-              style={{
-                height: 500,
-                marginTop: 20,
-                marginBottom: 20,
-                display: "flex",
-                justifyContent: "center",
-                alignItems: "center"
-              }}>
-              {selectedWidget.widget}
-            </div>
+            {selectedWidget.name === intl.formatMessage(newId) ? (
+              <this.props.onBuild />
+            ) : (
+              <div
+                style={{
+                  height: 500,
+                  marginTop: 20,
+                  marginBottom: 20,
+                  display: "flex",
+                  justifyContent: "center",
+                  alignItems: "center"
+                }}>
+                {selectedWidget.widget}
+              </div>
+            )}
           </Modal.Content>
           <Modal.Footer>
             <div className="btn-group">
               <button
                 className="btn btn-warning-outline"
-                style={{ marginRight: "20px" }}
-                onClick={() => onClose()}>
+                onClick={() => onClose()}
+                style={{ marginRight: "20px" }}>
                 <Msg {...messages.modal.cancel} />
               </button>
-              <button className="btn btn-confirmatory" onClick={this.confirm}>
+              <button
+                className="btn btn-confirmatory"
+                onClick={this.confirm}
+                style={{
+                  display:
+                    selectedWidget.name === intl.formatMessage(newId) ? "none" : "inline-block"
+                }}>
                 <Msg {...messages.modal.confirm} />
               </button>
             </div>
