@@ -10,7 +10,7 @@ import "react-resizable/css/styles.css"
 import styles from "./styles.css"
 import { injectIntl } from "react-intl"
 
-const MARGIN = 40
+const MARGIN = 10
 
 class WidgetError extends React.Component {
   state = { error: false }
@@ -95,6 +95,7 @@ export default class Dashboard extends React.Component {
       ...currentLayouts,
       [breakpoint]: [...(currentLayouts[breakpoint] || []), widget]
     }
+    this.props.onSave(layouts)
     this.setState({ layouts, add: null })
   }
 
@@ -174,21 +175,22 @@ export default class Dashboard extends React.Component {
 
     const unAddedWidgets = this.getUnaddedWidgets()
     const widgets = this.getWidgets()
-
     return (
       <UnMarginDiv size={margin}>
         {!free && (
-          <Actions
-            canBuild={onBuild}
-            onBuild={onBuild}
-            canAdd={unAddedWidgets.length > 0}
-            onSave={this.save}
-            onUndo={this.undoEdit}
-            onEdit={this.edit}
-            onAdd={this.add}
-            editable={edit}
-            saving={saving}
-          />
+          <div style={{ marginBottom: 40 }}>
+            <Actions
+              canBuild={onBuild}
+              onBuild={onBuild}
+              canAdd={unAddedWidgets.length > 0}
+              onSave={this.save}
+              onUndo={this.undoEdit}
+              onEdit={this.edit}
+              onAdd={this.add}
+              editable={edit}
+              saving={saving}
+            />
+          </div>
         )}
         {add && (
           <AddWidgetModal
@@ -196,6 +198,7 @@ export default class Dashboard extends React.Component {
             messages={messages}
             onAdd={this.addWidget}
             onClose={this.undoAdd}
+            onBuild={onBuild}
           />
         )}
         <ResponsiveReactGridLayout
