@@ -1,23 +1,32 @@
-import React from 'react'
-import Websocket from 'react-websocket'
-import { FormattedMessage } from 'react-intl'
-import { Pagination } from 'aurigauikit/components/SimpleTable'
-import NotificationService2 from './notification-service-new'
-import { withRouter } from 'react-router-dom'
-import styles from './style.less'
-import Popover from 'aurigauikit/components/Popover'
+import React, { createRef } from "react"
+import { FormattedMessage } from "react-intl"
+import { Pagination } from "aurigauikit/components/SimpleTable"
+import NotificationService2 from "./notification-service-new"
+import { withRouter } from "react-router-dom"
+import styles from "./style.less"
+import Popover from "aurigauikit/components/Popover"
+import { NativeEventSource, EventSourcePolyfill } from "event-source-polyfill"
+
+const EventSource = NativeEventSource || EventSourcePolyfill
+// OR: may also need to set as global property
+window.EventSource = NativeEventSource || EventSourcePolyfill
+
+let sse = null
 
 const getURL = (
-  { hostname = location.hostname, port = 8080, ['context-path']: contextPath = '' },
+  { protocol, hostname = location.hostname, port = 8080, ["context-path"]: contextPath = "" },
   userCode,
   appCode
 ) =>
-  `ws://${hostname}:${port}${contextPath}/api/v1/notification/socket/${appCode}/${userCode
-    .split(' ')
-    .join('')}`
+  `${protocol ||
+    "http"}://${hostname}:${port}${contextPath}/api/v1/notification/sse/${appCode}/${userCode
+    .split(" ")
+    .join("")}`
 
 @withRouter
 class Notification extends React.Component {
+  notificationRef = createRef()
+
   constructor(props) {
     super(props)
     this.state = {
@@ -25,9 +34,23 @@ class Notification extends React.Component {
       unreadCount: 0,
       notifications: [],
       page: 1,
-      position: null,
+      position: null
     }
     this.service = new NotificationService2(props.frontend, props.backend)
+  }
+
+  componentDidMount() {
+    const { userCode, appCode, backend, customUrl } = this.props
+    let url =
+      backend === "string"
+        ? `/${backend}/api/v1/notification/sse/${appCode}/${userCode}`
+        : getURL(backend, userCode, appCode)
+    if (customUrl) url = customUrl
+    if (!sse) {
+      sse = new EventSource(url)
+      sse.onmessage = this.handleNotification
+    }
+    this.connectionOpen()
   }
 
   connectionOpen = async () => {
@@ -39,7 +62,7 @@ class Notification extends React.Component {
     this.setState({
       active: true,
       unreadCount,
-      notifications,
+      notifications
     })
   }
 
@@ -53,11 +76,11 @@ class Notification extends React.Component {
     })
   }
 
-  handleNotification = data => {
+  handleNotification = ({ data }) => {
     this.setState(prevState => ({
       active: true,
       unreadCount: prevState.unreadCount + 1,
-      notifications: [JSON.parse(data), ...prevState.notifications],
+      notifications: [JSON.parse(data), ...prevState.notifications]
     }))
   }
 
@@ -69,7 +92,7 @@ class Notification extends React.Component {
 
   onPageChange = page => {
     this.setState({
-      page,
+      page
     })
   }
 
@@ -77,7 +100,7 @@ class Notification extends React.Component {
     notification.read = !notification.read
     await this.service.readOrUnreadNotification(notification.id, notification.read)
     this.setState(prevState => ({
-      unreadCount: prevState.unreadCount + (notification.read ? -1 : 1),
+      unreadCount: prevState.unreadCount + (notification.read ? -1 : 1)
     }))
   }
 
@@ -87,9 +110,11 @@ class Notification extends React.Component {
     this.setState(
       prevState => ({
         unreadCount: prevState.unreadCount - 1,
-        position: null,
+        position: null
       }),
-      () => this.props.history.push(notification.reference)
+      () => {
+        if (notification.reference) this.props.history.push(notification.reference)
+      }
     )
   }
 
@@ -98,7 +123,7 @@ class Notification extends React.Component {
     this.setState(
       prevState => ({
         unreadCount: prevState.unreadCount - (!notification.read ? 1 : 0),
-        notifications: prevState.notifications.filter(n => n.id !== notification.id),
+        notifications: prevState.notifications.filter(n => n.id !== notification.id)
       }),
       () => {
         this.forceUpdate()
@@ -115,26 +140,26 @@ class Notification extends React.Component {
   render() {
     if (this.state.error)
       return (
-        <div style={{ display: 'inline-block' }}>
-          <span style={{ color: 'white' }}>
+        <div style={{ display: "inline-block" }}>
+          <span style={{ color: "white" }}>
             <i className="fa fa-bell-slash" style={{ marginLeft: 9 }} />
           </span>
         </div>
       )
-    const { userCode, appCode, backend, customUrl } = this.props
     const pageSize = 6
     const { notifications, page, position } = this.state
     const unreadCount = notifications.filter(n => !n.read).length
-    let webSocketUrl =
-      backend === 'string'
-        ? `/${backend}/api/v1/notification/${userCode}`
-        : getURL(backend, userCode, appCode)
-    if (customUrl) webSocketUrl = customUrl
-    const openPopover = e => this.openPopover({ left: e.pageX, top: e.pageY + 15 })
+    const openPopover = () => {
+      const rect = this.notificationRef.current.getBoundingClientRect()
+      this.openPopover({
+        left: (rect.x || rect.left) + rect.width / 2,
+        top: (rect.y || rect.top) + rect.height + 5
+      })
+    }
     const content = (
       <React.Fragment>
         {notifications && notifications.length > 0 && (
-          <div style={{ pointerEvents: 'auto', width: 350 }}>
+          <div style={{ pointerEvents: "auto", width: 350 }}>
             <ul className="Notification--ul">
               {notifications.slice(pageSize * (page - 1), pageSize * page).map(notification => {
                 return (
@@ -145,8 +170,8 @@ class Notification extends React.Component {
                     <i
                       className={`fa fa-circle greater-feedback Notification--check`}
                       style={{
-                        color: notification.read ? '#ccc' : '#2984C5',
-                        textShadow: notification.read ? null : '0px 0px 10px #2984C5',
+                        color: notification.read ? "#ccc" : "#2984C5",
+                        textShadow: notification.read ? null : "0px 0px 10px #2984C5"
                       }}
                       onClick={e => {
                         e.stopPropagation()
@@ -156,7 +181,7 @@ class Notification extends React.Component {
                     <div
                       className="Notification--content"
                       style={{
-                        fontWeight: notification.read ? 'normal' : 'bold',
+                        fontWeight: notification.read ? "normal" : "bold"
                       }}>
                       {notification.message}
                     </div>
@@ -176,7 +201,7 @@ class Notification extends React.Component {
                 style={{
                   padding: 15,
                   paddingBottom: 10,
-                  paddingTop: 0,
+                  paddingTop: 0
                 }}>
                 <Pagination
                   page={page}
@@ -191,12 +216,12 @@ class Notification extends React.Component {
           (notifications.length === 0 && (
             <div
               style={{
-                pointerEvents: 'auto',
+                pointerEvents: "auto",
                 fontWeight: 600,
                 fontSize: 16,
                 padding: 15,
                 width: 160,
-                color: 'rgb(233, 128, 54)',
+                color: "rgb(233, 128, 54)"
               }}>
               <FormattedMessage id="notification.emptyList" />
             </div>
@@ -204,23 +229,23 @@ class Notification extends React.Component {
       </React.Fragment>
     )
     return (
-      <div style={{ display: 'inline-block' }}>
+      <div style={{ display: "inline-block" }}>
         <span
           onClick={openPopover}
           style={{
-            color: unreadCount > 0 ? '#2984C5' : 'white',
-            textShadow: unreadCount > 0 ? '0px 0px 10px #2984C5' : null,
-            cursor: 'pointer',
+            color: unreadCount > 0 ? "#2984C5" : "white",
+            textShadow: unreadCount > 0 ? "0px 0px 10px #2984C5" : null,
+            cursor: "pointer"
           }}>
-          <i className="fa fa-bell" style={{ marginLeft: 9 }} />
+          <i className="fa fa-bell" style={{ marginLeft: 9 }} ref={this.notificationRef} />
           <span
             style={{
-              color: unreadCount > 0 ? 'white' : 'white',
-              textShadow: unreadCount > 0 ? '0px 0px 10px white' : null,
-              fontWeight: 'bold',
-              cursor: 'pointer',
-              position: 'absolute',
-              transform: 'translateX(-29px) translateY(-5px)',
+              color: unreadCount > 0 ? "white" : "white",
+              textShadow: unreadCount > 0 ? "0px 0px 10px white" : null,
+              fontWeight: "bold",
+              cursor: "pointer",
+              position: "absolute",
+              transform: "translateX(-29px) translateY(-5px)"
             }}>
             {unreadCount > 0 && unreadCount}
           </span>
@@ -228,14 +253,6 @@ class Notification extends React.Component {
         <Popover key="popover" position={position} onClose={this.closePopover}>
           {content}
         </Popover>
-        <div style={{ display: 'none' }}>
-          <Websocket
-            url={webSocketUrl}
-            onMessage={this.handleNotification}
-            onOpen={this.connectionOpen}
-            onClose={this.connectionClose}
-          />
-        </div>
       </div>
     )
   }
