@@ -1,21 +1,10 @@
-// @flow
 import Ajax from "aurigauikit/ajax"
-
-type Notification = {
-  id: number,
-  type: string,
-  refDate: number,
-  message: string,
-  ref: string,
-  read: boolean
-}
-
 class NotificationService {
   constructor(frontend, backend) {
     this.ajax = Ajax(frontend)(backend)
   }
 
-  fetchNotification = (): Notification[] => {
+  fetchNotification = () => {
     return this.ajax.get("notification")
   }
 
