@@ -1,479 +1,285 @@
-import React from "react"
-import { FormattedMessage, injectIntl } from "react-intl"
+import React, { useState, useEffect, useMemo, useRef } from 'react'
+import { FormattedMessage, injectIntl } from 'react-intl'
+import SimpleTable from '../../SimpleTable/SimpleTable'
+import Tree from '../../Tree/Tree'
+import Modal from '../../Modal'
+import Radio from 'aurigauikit/components/Radio'
+import Spinner from 'aurigauikit/components/Spinner'
+import messages from '../messages'
+import OrganizationalUnit from '../OrganizationalUnit'
+import './organizational-unit-modal.less'
 
-import isEqual from "lodash.isequal"
+const dataComparator = (e1, e2) => e1 && e2 && e1.type === e2.type && e1.id === e2.id
+const isEqual = (a, b) => a.reduce((equal, a, i) => equal && dataComparator(a, b[i]), true)
 
-import SimpleTable from "../../SimpleTable/SimpleTable"
-import Tree from "../../Tree/Tree"
-import Modal from "../../Modal"
-import Radio from "aurigauikit/components/Radio"
-import Spinner from "aurigauikit/components/Spinner"
-import Checkbox from "aurigauikit/components/Checkbox"
-import { Grid, Div } from "aurigauikit/components/Grid"
-import StaticPanel from "aurigauikit/components/StaticPanel"
-import GroupingTable from "./GroupingTable"
-import "./organizational-unit-modal.less"
-import messages from "../messages"
+const useOuTree = (datasource, intl) => {
+  const currentRequest = useRef(null)
+  const [loading, setLoading] = useState(false)
+  const [page, onPageChange] = useState(1)
 
-class OrganizationalUnitModal extends React.Component {
-  constructor(props) {
-    super(props)
-    this.state = {
-      tree: [],
-      openPath: null,
-      grouping: false,
-      selectedPath: null,
-      loadingPath: null,
-      selectedElements: this.props.selectedElements
-        ? this.props.selectedElements
-        : this.props.defaultSelection
-          ? [this.props.defaultSelection]
-          : [],
-      tablePage: 1,
-      selectedOption: this.props.radioOptions ? this.props.radioOptions[0] : null,
-      datasource: this.props.datasource
+  const [tree, setTree] = useState({
+    data: [],
+    openPath: null,
+    selectedPath: null,
+    loadingPath: null,
+  })
+
+  const element = useMemo(() => search(tree.data, tree.selectedPath), [
+    tree.data,
+    tree.selectedPath,
+  ])
+
+  const table = useMemo(() => {
+    if (!element || !element.table) {
+      return {
+        caption: null,
+        headers: [],
+        columns: [],
+        data: [],
+        page,
+        onPageChange,
+      }
     }
-  }
+    const caption = `${intl.formatMessage(
+      messages.type[element.node.getType()]
+    )} - ${element.node.getDescription()} (${element.node.getCode()})`
 
-  componentWillReceiveProps(props) {
-    if (
-      this.props.selectedElements &&
-      (!isEqual(props.selectedElements, this.props.selectedElements) ||
-        !isEqual(props.selectedElements, this.state.selectedElements))
-    ) {
-      this.setState({ selectedElements: props.selectedElements })
-    }
-    if (!props.radioOptions && props.datasource !== this.state.datasource) {
-      this.setState({ datasource: props.datasource })
-    }
-    if (this.props.show !== props.show && props.show) {
-      this.setState({ tree: null }, () => this.componentDidMount())
-    }
-  }
+    const headers = element.table.headers.map(header =>
+      intl.formatMessage(messages.modal.headers[header])
+    )
 
-  onConfirm() {
-    this.props.onSelectionConfirmed(this.state.selectedElements)
-  }
-
-  onAbort() {
-    if (this.props.onSelectionAborted) this.props.onSelectionAborted()
-  }
-
-  onSelect(selectedElements) {
-    this.setState({ selectedElements: selectedElements })
-  }
-
-  onReset() {
-    this.setState({
-      selectedElements: this.props.defaultSelection ? [this.props.defaultSelection] : []
+    const columns = element.table.columns.map(column => {
+      if (column === 'type')
+        return { content: json => intl.formatMessage(messages.type[json.type]) }
+      else return column
     })
-  }
+    return { caption, headers, columns, data: element.table.rows, page, onPageChange }
+  }, [element])
 
-  onPageChange(page) {
-    this.setState({ tablePage: page })
-  }
-
-  componentDidMount() {
-    const { radioOptions } = this.props
-    if (radioOptions) {
-      this.setState(
-        {
-          selectedOption: radioOptions[0],
-          datasource: radioOptions[0].datasource,
-          loading: true
-        },
-        () => this.fetchInitialData()
-      )
-    } else this.setState({ loading: true }, () => this.fetchInitialData())
-  }
-
-  async fetchInitialData() {
-    const tree = await this.state.datasource.getNodes()
-    const content = await this.fetchContent({ path: "0", node: tree[0] })
-    this.setState({
-      tree: tree.map((element, index) => ({
-        node: element,
-        table: index === 0 ? content.table : null,
-        path: index + "",
-        children: index === 0 ? content.children : []
-      })),
-      openPath: "0",
-      selectedPath: "0",
-      loading: false,
-      forceLoading: false
-    })
-  }
-
-  async fetchContent({ node, path }) {
-    const [nodes, table] = await Promise.all([
-      this.state.datasource.getNodes(node),
-      this.state.datasource.getTable(node)
-    ])
-
+  async function fetchContent({ node, path }) {
+    const [nodes, table] = await Promise.all([datasource.getNodes(node), datasource.getTable(node)])
     const children = nodes.map((child, index) => ({
       node: child,
-      path: [path, index].join("-"),
-      children: []
+      path: [path, index].join('-'),
+      children: [],
     }))
-
-    const rows = table.getRowCount
+    const rows = (table.getRowCount
       ? Array.from(Array(table.getRowCount())).map((_, i) => table.getRow(i))
       : []
-
+    ).map((a, index) => ({ ...a, path: [path, index].join('-') }))
     const headers = table.getHeaders ? table.getHeaders() : []
     const columns = table.getColumns ? table.getColumns() : []
-
-    return {
-      children,
-      table: { rows, headers, columns }
-    }
+    return { children, table: { rows, headers, columns } }
   }
 
-  async updateTree(element) {
-    const content = await this.fetchContent(element)
-    const updatedTree = updateTree(this.state.tree, element.path, content)
-    this.setState({ tree: updatedTree })
-  }
-
-  onTreeClick(treeState) {
-    const element = search(this.state.tree, treeState.selectedPath)
-    const newState = {
+  async function update(treeState) {
+    const element = search(tree.data, treeState.selectedPath)
+    const newTree = {
       openPath: treeState.openPath,
       selectedPath: treeState.selectedPath,
       loadingPath: null,
-      tablePage: 1
     }
-    if (element.table) this.setState(newState)
+    if (element.table && element.table.rows.length > 0) setTree(tree => ({ ...tree, ...newTree }))
     else {
-      this.setState({ loadingPath: treeState.selectedPath })
-      this.updateTree(element)
-        .then(() => this.setState(newState))
-        .catch(e => {
-          console.error("ERROR", e) //eslint-disable-line no-console
-          this.setState({
-            loadingPath: null,
-            tablePage: 1
-          })
-        })
+      const request = (currentRequest.current = new Date().valueOf())
+      setTree(tree => ({ ...tree, loadingPath: treeState.selectedPath }))
+      const content = await fetchContent(element)
+      const updatedTree = updateTree(tree.data, element.path, content)
+      if (request === currentRequest.current) {
+        setTree(tree => ({ ...tree, ...newTree, data: updatedTree }))
+        table.onPageChange(1)
+      }
     }
   }
 
-  getColumns(columns) {
-    return columns.map(column => {
-      if (column === "type")
-        return {
-          content: json => this.props.intl.formatMessage(messages.type[json.type])
-        }
-      else return column
-    })
+  async function fetch() {
+    const request = (currentRequest.current = new Date().valueOf())
+    setLoading(true)
+    const tree = await datasource.getNodes()
+    const content = await fetchContent({ path: '0', node: tree[0] })
+    if (request === currentRequest.current) {
+      setTree(state => ({
+        ...state,
+        data: tree.map((element, index) => ({
+          node: element,
+          table: index === 0 ? content.table : null,
+          path: index + '',
+          children: index === 0 ? content.children : [],
+        })),
+      }))
+      table.onPageChange(1)
+      setLoading(false)
+    }
   }
 
-  onGroup = () => this.setState({ grouping: true })
-  onGroupUndo = () => this.setState({ grouping: false })
-  static defaultProps = {
-    dataComparator: (e1, e2) => e1 && e2 && e1.type === e2.type && e1.id === e2.id,
-    canSelect: () => true
-  }
-  render() {
-    const element = search(this.state.tree, this.state.selectedPath)
-    if (this.state.selectedPath && element && element.table) var table = element.table // eslint-disable-line no-var
-    const hasSelected = !(
-      this.state.selectedElements &&
-      this.state.selectedElements.length === 0 + (this.props.defaultSelection ? 1 : 0)
-    )
-    const {
-      show,
-      canSelect,
-      defaultSelection,
-      dataComparator,
-      radioOptions,
-      groupable,
-      intl
-    } = this.props
-    return (
-      <Modal
-        style={{ width: "70%", minHeight: !this.state.grouping && "600px" }}
-        show={show}
-        onClose={() => this.onAbort()}>
-        <Modal.Header
-          title={
-            <span>
-              <FormattedMessage {...messages.general.title} />
-              {this.state.grouping && <span> - New Group </span>}
-            </span>
-          }>
-          <div style={{ marginRight: -10, marginTop: -40 }}>
-            {!this.state.grouping && (
-              <ResetButton disabled={!hasSelected} onClick={() => this.onReset()} />
-            )}
-            {groupable &&
-              !this.state.grouping && (
-                <GroupButton disabled={!hasSelected} onClick={() => this.onGroup()} />
-              )}
-          </div>
-        </Modal.Header>
-        {!this.state.grouping && (
-          <Modal.Content>
-            {radioOptions && (
-              <div style={{ display: "inline-block", width: "100%" }}>
-                {radioOptions.map((radioOption, i) => (
-                  <Radio
-                    key={i}
-                    style={{
-                      float: "left",
-                      paddingRight: 20,
-                      paddingBottom: 10
-                    }}
-                    isChecked={isEqual(this.state.selectedOption, radioOption)}
-                    onChange={checked => {
-                      if (checked)
-                        this.setState(
-                          {
-                            selectedOption: radioOption,
-                            datasource: radioOption.datasource,
-                            loading: true,
-                            forceLoading: true
-                          },
-                          () => this.fetchInitialData()
-                        )
-                    }}>
-                    {radioOption.i18nLabel ? (
-                      <FormattedMessage id={radioOption.i18nLabel} />
-                    ) : (
-                      radioOption.message
-                    )}
-                  </Radio>
-                ))}
-              </div>
-            )}
-            <Body>
-              {((!this.state.tree && this.state.loading) || this.state.forceLoading) && (
-                <Spinner target="org-modal-body" config={{ top: "calc(22vh)" }} />
-              )}
-              <Left>
-                <Tree
-                  onElementClick={state => this.onTreeClick(state)}
-                  openPath={this.state.openPath}
-                  selectedPath={this.state.selectedPath}
-                  willDisplay={node => `${node.getDescription()} (${node.getCode()})`}
-                  loadingPath={this.state.loadingPath}
-                  canView={this.state.selectedOption && this.state.selectedOption.canView}
-                  data={this.state.tree}
-                />
-              </Left>
-              <Right>
-                <h3 style={{ margin: 0, marginBottom: 10, fontSize: 15 }}>
-                  {table &&
-                    `${intl.formatMessage(
-                      messages.type[element.node.getType()]
-                    )} - ${element.node.getDescription()} (${element.node.getCode()})`}
-                </h3>
-                {table && (
-                  <SimpleTable
-                    selectable
-                    pageable
-                    pageSize={8}
-                    data={table.rows}
-                    onSelectionChange={selectedElements => this.onSelect(selectedElements)}
-                    selectedRows={this.state.selectedElements}
-                    headers={table.headers.map(
-                      header => intl.formatMessage(messages.modal.headers[header])
-                    )}
-                    canSelect={element =>
-                      this.state.selectedOption
-                        ? this.state.selectedOption.canSelect(element)
-                        : canSelect(element)
-                    }
-                    onPageChange={page => this.onPageChange(page)}
-                    page={this.state.tablePage}
-                    defaultSelection={[defaultSelection]}
-                    dataComparator={dataComparator}
-                    columns={this.getColumns(table.columns)}
-                  />
-                )}
-              </Right>
-            </Body>
-          </Modal.Content>
-        )}
-        {this.state.grouping && (
-          <Modal.Content>
-            <section style={{ marginTop: 20 }}>
-              <Grid>
-                <Div col="1-4">
-                  <label>
-                    {/*<FormattedMessage id="branch-groups-page.group-modal.insert-code" />*/}
-                    Description
-                  </label>
-                  <input
-                    style={{ width: "100%" }}
-                    type="text"
-                    onChange={e => this.setState({ groupCode: e.target.value })}
-                  />
-                </Div>
-                <Div col="9-12">
-                  <label>
-                    {/*<FormattedMessage id="branch-groups-page.group-modal.insert-desc" />*/}
-                    Description
-                  </label>
-                  <input
-                    style={{ width: "100%" }}
-                    type="text"
-                    onChange={e => this.setState({ groupDescription: e.target.value })}
-                  />
-                </Div>
-              </Grid>
-            </section>
-            <section style={{ marginBottom: 60 }}>
-              <GroupingTable
-                data={this.state.selectedElements}
-                columns={this.getColumns(table.columns)}
-                headers={table.headers.map(header =>
-                  intl.formatMessage(messages.modal.headers[header])
-                )}
-              />
-            </section>
-          </Modal.Content>
-        )}
-        <Modal.Footer>
-          <FooterButtons>
-            <Selection elements={this.state.selectedElements} />
-            {!this.state.grouping && <AbortButton onClick={() => this.onAbort()} />}
-            {!this.state.grouping && (
-              <ConfirmButton
-                disabled={!hasSelected && !this.props.defaultSelection}
-                onClick={() => this.onConfirm()}
-              />
-            )}
-            {this.state.grouping && (
-              <Checkbox
-                style={{ float: "left", marginLeft: 80 }}
-                isChecked={this.state.notPublic}
-                onChange={notPublic => this.setState({ notPublic: notPublic })}>
-                <FormattedMessage {...messages.modal.general.notPublic} />
-              </Checkbox>
-            )}
-            {this.state.grouping && <GroupUndoButton onClick={() => this.onGroupUndo()} />}
-            {this.state.grouping && (
-              <GroupConfirmButton
-                disabled={!hasSelected && !this.props.defaultSelection}
-                onClick={() => this.onConfirm()}
-              />
-            )}
-          </FooterButtons>
-        </Modal.Footer>
-      </Modal>
-    )
-  }
+  return { ...tree, element, loading, update, fetch, table }
 }
 
-const Body = ({ children }) => (
-  <div id="org-modal-body" className="grid organizational-unit-grid">
-    {children}
+const OuModal = ({
+  selectedElements,
+  defaultSelection,
+  datasource,
+  onAbort,
+  onSelect,
+  onRemove,
+  canSelect,
+  canView,
+  canRemove,
+  radioOptions,
+  intl,
+  onSelectionConfirmed = () => {},
+  onSelectionAborted = () => {},
+}) => {
+  const intitialData = useMemo(() => selectedElements, [])
+  const touched = useMemo(() => !isEqual(selectedElements, intitialData), [selectedElements])
+  const [option, setOption] = useState(null)
+  const currentDatasource = option ? option.datasource : datasource
+
+  const tree = useOuTree(currentDatasource, intl)
+
+  useEffect(() => {
+    if (radioOptions && !option) {
+      const [firstOption] = radioOptions
+      setOption(firstOption)
+    } else tree.fetch()
+  }, [option])
+
+  useEffect(() => {
+    if (defaultSelection && selectedElements.length === 0) onSelect([defaultSelection])
+  }, [])
+
+  const canReset = useMemo(() => {
+    if (defaultSelection && selectedElements.find(a => dataComparator(a, defaultSelection)))
+      return selectedElements.length > 1
+    return selectedElements.length > 0
+  })
+
+  const selectElements = elements => {
+    const newElements = elements.filter(n => !selectedElements.find(e => dataComparator(e, n)))
+    const getParents = path => {
+      if (path === '0') return []
+      const paths = (path || '0').split('-')
+      paths.pop()
+      const parentPath = paths.join('-')
+      const parent = elements.find(e => (e.path || '0') === parentPath)
+      return [...getParents(parentPath), parent]
+    }
+    const parents = newElements
+      .reduce((parents, e) => [...parents, ...getParents(e.path)], [])
+      .filter(Boolean)
+    const children = newElements
+      .reduce((children, e) => {
+        return [
+          ...children,
+          ...elements.filter(n => n.path !== e.path && (n.path || '0').startsWith(e.path || '0')),
+        ]
+      }, [])
+      .filter(Boolean)
+    const filtered = elements.filter(e => {
+      return !(parents.find(a => dataComparator(a, e)) || children.find(a => dataComparator(a, e)))
+    })
+    onSelect(filtered.length === 0 ? [defaultSelection] : filtered)
+  }
+
+  return (
+    <Modal
+      show
+      style={{ width: '70%' }}
+      onClose={() => {
+        onSelect(intitialData)
+        onAbort()
+      }}>
+      <Modal.Header
+        title={
+          <span>
+            <FormattedMessage {...messages.general.title} />
+          </span>
+        }>
+        <div style={{ marginRight: -10, marginTop: -40 }}>
+          <ResetButton disabled={!canReset} onClick={() => onSelect([defaultSelection])} />
+        </div>
+      </Modal.Header>
+      <Modal.Content>
+        {radioOptions && option && (
+          <DatasourceSelector
+            datasources={radioOptions}
+            datasource={currentDatasource}
+            onChange={setOption}
+          />
+        )}
+        <div id="org-modal-body" className="grid organizational-unit-grid">
+          {tree.loading && <Spinner target="org-modal-body" config={{ top: '220px' }} />}
+          <div className="col-7-12 organizational-unit-tree">
+            <Tree
+              onElementClick={tree.update}
+              openPath={tree.openPath}
+              selectedPath={tree.selectedPath}
+              willDisplay={node => `${node.getDescription()} (${node.getCode()})`}
+              loadingPath={tree.loadingPath}
+              canView={(option ? option.canView : canView) || canView}
+              data={tree.data}
+            />
+          </div>
+          <div className="col-5-12 organizational-unit-table">
+            {tree.element && tree.element.table && (
+              <SimpleTable
+                selectable
+                pageable
+                pageSize={8}
+                canSelect={(option ? option.canSelect : canSelect) || canSelect}
+                onSelectionChange={selectElements}
+                selectedRows={selectedElements}
+                dataComparator={dataComparator}
+                {...tree.table}
+              />
+            )}
+          </div>
+        </div>
+        <div style={{ float: 'left', marginTop: 20 }}>
+          <OrganizationalUnit data={selectedElements} onRemove={onRemove} canRemove={canRemove} />
+        </div>
+        <div style={{ float: 'right', marginTop: 20 }}>
+          <AbortButton
+            onClick={() => {
+              onSelect(intitialData)
+              onAbort()
+              onSelectionAborted()
+            }}
+          />
+          <ConfirmButton
+            disabled={!touched}
+            onClick={() => {
+              onAbort()
+              onSelectionConfirmed(selectedElements)
+            }}
+          />
+        </div>
+      </Modal.Content>
+      <Modal.Footer />
+    </Modal>
+  )
+}
+
+const DatasourceSelector = ({ datasource, datasources, onChange }) => (
+  <div style={{ display: 'inline-block', width: '100%' }}>
+    {datasources.map((option, i) => (
+      <Radio
+        key={i}
+        style={{ float: 'left', paddingRight: 20, paddingBottom: 10 }}
+        isChecked={datasource === option.datasource}
+        onChange={checked => checked && onChange(option)}>
+        {option.i18nLabel ? <FormattedMessage id={option.i18nLabel} /> : option.label}
+      </Radio>
+    ))}
   </div>
 )
-const Left = ({ children }) => <div className="col-7-12 organizational-unit-tree"> {children} </div>
-const Right = ({ children }) => (
-  <div className="col-5-12 organizational-unit-table"> {children} </div>
-)
-const FooterButtons = ({ children }) => <div className="btn-group"> {children} </div>
-
-const sameType = (type, ous) => ous.filter(ou => ou.type === type)
-const allTypes = ous => ous.map(ou => ou.type).filter((c, i, a) => a.indexOf(c) === i)
-
-const Selection = injectIntl(({ elements, intl }) => {
-  const Panel = ({ children }) => (
-    <StaticPanel
-      style={{
-        float: "left",
-        marginLeft: 80,
-        padding: 0,
-        paddingRight: 10,
-        paddingLeft: 10
-      }}>
-      {children}
-    </StaticPanel>
-  )
-  if (elements && elements.length === 0) {
-    return (
-      <Panel>
-        <em>No elements selected</em>
-      </Panel>
-    )
-  }
-  const types = allTypes(elements)
-  return (
-    <Panel>
-      <em>Selection</em>{" "}
-      {types.map((type, i) => {
-        const sameElements = sameType(type, elements)
-        return (
-          <span key={i}>
-            <b>{intl.formatMessage(messages.type[type])}</b>: {sameElements && sameElements.length}
-            {types && i < types.length - 1 && ", "}
-          </span>
-        )
-      })}
-    </Panel>
-  )
-})
-
-const GroupButton = ({ disabled, onClick }) => {
-  const style = {
-    float: "right",
-    font: '13.3333px "Open Sans", sans-serif',
-    fontWeight: "bold",
-    margin: 10
-  }
-  return (
-    <button
-      disabled={disabled}
-      className="btn btn-confirmatory-outline"
-      style={style}
-      onClick={() => onClick()}>
-      <i
-        className="fa fa-compress"
-        style={{
-          textAlign: "center",
-          padding: "2px",
-          lineHeight: "32px",
-          cursor: "pointer"
-        }}
-      />
-      <FormattedMessage {...messages.modal.groupSelection} />
-    </button>
-  )
-}
-
-const UndoButton = ({ onClick }) => {
-  const style = {
-    float: "right",
-    font: '13.3333px "Open Sans", sans-serif',
-    fontWeight: "bold",
-    margin: 10
-  }
-  return (
-    <button className="btn btn-warning-outline" style={style} onClick={() => onClick()}>
-      <i
-        className="fa fa-compress"
-        style={{
-          textAlign: "center",
-          padding: "2px",
-          lineHeight: "32px",
-          cursor: "pointer"
-        }}
-      />
-      {/*<FormattedMessage id="organizational-unit-modal.group-selection" />*/}
-      Add more
-    </button>
-  )
-}
 
 const ResetButton = ({ disabled, onClick }) => {
   const style = {
-    float: "right",
+    float: 'right',
     font: '13.3333px "Open Sans", sans-serif',
-    fontWeight: "bold",
-    margin: 10
+    fontWeight: 'bold',
+    margin: 10,
   }
   return (
     <button
@@ -489,7 +295,7 @@ const ResetButton = ({ disabled, onClick }) => {
 const AbortButton = ({ onClick }) => (
   <button
     className="btn btn-warning-outline"
-    style={{ marginRight: "20px" }}
+    style={{ marginRight: '20px' }}
     onClick={() => onClick()}>
     <FormattedMessage {...messages.modal.general.cancel} />
   </button>
@@ -498,23 +304,6 @@ const AbortButton = ({ onClick }) => (
 const ConfirmButton = ({ disabled, onClick }) => (
   <button disabled={disabled} className="btn btn-confirmatory" onClick={() => onClick()}>
     <FormattedMessage {...messages.modal.general.select} />
-  </button>
-)
-
-const GroupUndoButton = ({ onClick }) => (
-  <button
-    className="btn btn-warning-outline"
-    style={{ marginRight: "20px" }}
-    onClick={() => onClick()}>
-    {/*<FormattedMessage id="organizational-unit-modal.cancel" />*/}
-    Undo
-  </button>
-)
-
-const GroupConfirmButton = ({ disabled, onClick }) => (
-  <button disabled={disabled} className="btn btn-confirmatory" onClick={() => onClick()}>
-    {/*<FormattedMessage id="organizational-unit-modal.select" />*/}
-    Save group
   </button>
 )
 
@@ -537,4 +326,4 @@ const updateTree = (tree, path, content) => {
   })
 }
 
-export default injectIntl(OrganizationalUnitModal)
+export default injectIntl(OuModal)
