@@ -70,22 +70,21 @@ class Table extends React.Component {
     const hasMenu = this.props.menu.items.length > 0
     return (
       <div style={{ width: "100%", overflowX: "auto", overflowY: "hidden", ...this.props.style }}>
-        {this.props.editable &&
-          this.props.cellEditor.context && (
-            <CellEditor
-              isOpen={this.props.cellEditor.isOpen}
-              title={this.props.cellEditor.context.title}
-              position={this.props.cellEditor.context.position}
-              validations={this.props.validations}
-              colIndex={this.props.cellEditor.context.colIndex}
-              rowIndex={this.props.cellEditor.context.rowIndex}
-              onEditAborted={() => this.props.onEditAborted()}
-              onEditConfirmed={(rowIndex, colIndex, value) =>
-                this.onCellEditConfirmed(rowIndex, colIndex, value)
-              }
-              value={this.props.cellEditor.context.value}
-            />
-          )}
+        {this.props.editable && this.props.cellEditor.context && (
+          <CellEditor
+            isOpen={this.props.cellEditor.isOpen}
+            title={this.props.cellEditor.context.title}
+            position={this.props.cellEditor.context.position}
+            validations={this.props.validations}
+            colIndex={this.props.cellEditor.context.colIndex}
+            rowIndex={this.props.cellEditor.context.rowIndex}
+            onEditAborted={() => this.props.onEditAborted()}
+            onEditConfirmed={(rowIndex, colIndex, value) =>
+              this.onCellEditConfirmed(rowIndex, colIndex, value)
+            }
+            value={this.props.cellEditor.context.value}
+          />
+        )}
         {hasMenu && (
           <ContextMenu
             items={this.props.menu.items}
@@ -94,12 +93,11 @@ class Table extends React.Component {
           />
         )}
         <table className={`table table-striped ${this.props.className}`}>
-          {this.props.caption &&
-            this.props.caption != "" && (
-              <caption style={{ pointerEvents: "none" }}>
-                <div>{this.props.caption}</div>
-              </caption>
-            )}
+          {this.props.caption && this.props.caption != "" && (
+            <caption style={{ pointerEvents: "none" }}>
+              <div>{this.props.caption}</div>
+            </caption>
+          )}
           <Head
             data={this.props.data}
             index={this.props.pageSize * (this.props.page - 1)}
@@ -147,14 +145,13 @@ class Table extends React.Component {
             loadingState={this.props.loadingState}
             loading={this.props.loading}
           />
-          {this.props.footers.length > 0 &&
-            this.props.loading && (
-              <tfoot>
-                <tr>
-                  <td colSpan={this.props.columns.length + 1} />
-                </tr>
-              </tfoot>
-            )}
+          {this.props.footers.length > 0 && this.props.loading && (
+            <tfoot>
+              <tr>
+                <td colSpan={this.props.columns.length + 1} />
+              </tr>
+            </tfoot>
+          )}
           {this.props.footers.length > 0 && (
             <Foot
               data={this.props.data}
