@@ -114,7 +114,7 @@ i18n.loadMessages("en", {
 
   "organizational-unit-select.type.PRIVATE": "Private",
   "organizational-unit-select.type.PUBLIC": "Public",
-  "organizational-unit-select.type.ASSET_GROUP": "Asset group",
+  "organizational-unit-select.type.ASSET_GROUP": "Groups",
 
   "notification.emptyList": "No notifications",
 
@@ -234,5 +234,7 @@ i18n.loadMessages("en", {
   "report.schedulation-table.report-name.head-table": "Schedulation name",
   "report.new-schedulation-form.report-duration": "Duration (days)",
   "report.new-schedulation-form.report-previous-month": "Previous Month",
-  "report-list-page.msg.pdf-view-error": "Error loading report"
+  "report-list-page.msg.pdf-view-error": "Error loading report",
+
+  "dashboard-page.add-widget.new": "New"
 })
